Adding composite parallelization to get_optical_data.py

+++ bin/get_optical_data.py
@@ -3,11 +3,12 @@ Downloads sentinel2 data using Sinergise API and creates composites for each spe
 locally and uploaded to s3 along with the Sentinel 2 files used to create the composite.
 """
 import argparse
+import multiprocessing
 import os
+import tqdm
 import subprocess
 from typing import List
 
-import boto3
 import numpy as np
 import yaml
 
@@ -23,10 +24,29 @@ from src.utilities.config_reader import CONFIG
 
 # TODO: Start thinking about file structure in s3 so we can hash each request configuration
 
+def _composite_task(task_args: argparse.Namespace):
+    multi_band_file_path = imaging.create_composite(
+        task_args.district_s2_dir,
+        task_args.district_composite_dir,
+        task_args.coord,
+        ['B02', 'B03', 'B04'],
+        np.float32,
+        task_args.slices,
+        task_args.n_cores > 1)
+
+    s3_session = initialize_s3(CONFIG.AWS.BUCKET)
+
+    upload_to_s3(s3_session, multi_band_file_path,
+                 os.path.join('composites', task_args.region, task_args.district,
+                              os.path.basename(multi_band_file_path)),
+                 bucket=task_args.s3_bucket)
+
+    return None
+
 
 def get_optical_data(sentinel_2_dir: str, composite_dir: str, bands: List[str], buffer: float, slices: int,
                      s3_bucket: str, requested_regions: List[str] = None, requested_districts: List[str] = None,
-                     keep_s2_dir: bool = False):
+                     upload_s2_dir: bool = False):
     s3_session = initialize_s3(CONFIG.AWS.BUCKET)
 
     with open(REGION_FILE_PATH, 'r') as file:
@@ -76,35 +96,54 @@ def get_optical_data(sentinel_2_dir: str, composite_dir: str, bands: List[str], 
             # create composite
             print('--------------------------------')
             print('creating composites')
+            n_cores = multiprocessing.cpu_count() - 1  # TODO: Make this a parameter
+            args = []
             for coord in os.listdir(district_s2_dir):
                 if coord.startswith("."):
                     continue
+                args.append(
+                    argparse.Namespace(
+                        district_s2_dir=district_s2_dir,
+                        district_composite_dir=district_composite_dir,
+                        coord=coord,
+                        slices=slices,
+                        n_cores=n_cores,
+                        s3_bucket=s3_bucket,
+                        region=region,
+                        district=district
+                    )
+                )
 
-                print(f'Creating composite for {region}/{district}/{coord}...')
-                multi_band_file_path = imaging.create_composite(district_s2_dir, district_composite_dir, coord, bands,
-                                                                np.float32,
-                                                                slices, False)
-                upload_to_s3(s3_session, multi_band_file_path,
-                             os.path.join('composites', region, district, os.path.basename(multi_band_file_path)),
-                             bucket=s3_bucket)
+            print('Building composites...')
+
+            # if n_cores == 1:
+            #     print('\tNot using multiprocessing...')
+            #     for arg in tqdm(args, total=len(args), desc="Sequential...", leave=True):
+            #         _composite_task(arg)
+            # else:
+            print('\tUsing multiprocessing...')
+            with multiprocessing.Pool(n_cores) as pool:
+                for _ in tqdm(pool.imap_unordered(_composite_task, args)):
+                    pass
 
             # Compress sentinel2
-            print('-----------------------------')
-            print('Compressing raw s2')
-            tar_file = os.path.join(B2P_DIR, 'data', 'sentinel2', f's2_{region}_{district}.tar.gz')
-            tar_cmd = f'tar -czvf {tar_file}'
-            if not keep_s2_dir:
+            if upload_s2_dir:
+                print('-----------------------------')
+                print('Compressing raw s2')
+                tar_file = os.path.join(B2P_DIR, 'data', 'sentinel2', f's2_{region}_{district}.tar.gz')
+                tar_cmd = f'tar -czvf {tar_file}'
+
                 tar_cmd += f' --remove-files {district_s2_dir}'
-            process = subprocess.Popen(tar_cmd.split(), shell=False)
-            process.communicate()
+                process = subprocess.Popen(tar_cmd.split(), shell=False)
+                process.communicate()
 
-            # Upload raw s2 to s3
-            upload_to_s3(s3_session, tar_file, os.path.join('sentinel2_raw', os.path.basename(tar_file)),
-                         bucket=s3_bucket)
+                # Upload raw s2 to s3
+                upload_to_s3(s3_session, tar_file, os.path.join('sentinel2_raw', os.path.basename(tar_file)),
+                             bucket=s3_bucket)
 
-            # remove raw data from this vm
-            os.remove(tar_file)
-            print('============================================\n\n')
+                # remove raw data from this vm
+                os.remove(tar_file)
+                print('============================================\n\n')
 
 
 if __name__ == "__main__":
@@ -125,7 +164,7 @@ if __name__ == "__main__":
                         default=os.path.join(B2P_DIR, 'data', 'composites'),
                         help='Path to the base directory where composites will be written to. Set to data/sentinel2 by'
                              ' default')
-    parser.add_argument('--keep_s2', '-s2', action='store_true',
+    parser.add_argument('--upload_s2', '-s2', action='store_true',
                         help='If specified then the s2 images will not be removed after being uploaded to s3 and will '
                              'be kept locally')
     parser.add_argument('--bands', '-b', required=False, nargs='+', type=str, default=['B02', 'B03', 'B04'],
@@ -142,4 +181,4 @@ if __name__ == "__main__":
 
     get_optical_data(requested_regions=args.regions, requested_districts=args.districts, sentinel_2_dir=args.s2_dir,
                      bands=args.bands, buffer=args.buffer, slices=args.slices, s3_bucket=args.s3_bucket,
-                     composite_dir=args.composite_dir, keep_s2_dir=args.keep_s2)
+                     composite_dir=args.composite_dir, upload_s2_dir=args.upload_s2)
